# pkg/apiserver/docs/swagger.yaml
definitions:
...
  api.DiskReqBody:
    properties:
      reserve:
        type: boolean
    type: object
...
  api.LocalDiskInfo:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      availableCapacityBytes:
        description: AvailableCapacityBytes 可用容量
        type: integer
      diskPathShort:
        description: diskPathShort 磁盘路径简写
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      localStoragePooLName:
        description: LocalStoragePooLName 存储池名称
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      spec:
        $ref: '#/definitions/v1alpha1.LocalDiskSpec'
      status:
        $ref: '#/definitions/v1alpha1.LocalDiskStatus'
      totalCapacityBytes:
        description: TotalCapacityBytes 总容量
        type: integer
    type: object
...
  api.LocalDiskListByNode:
    properties:
      diskPathShort:
        description: diskPathShort 磁盘路径简写
        type: string
      items:
        description: localDisks 节点磁盘列表
        items:
          $ref: '#/definitions/api.LocalDiskInfo'
        type: array
      nodeName:
        description: nodeName 节点名称
        type: string
      pagination:
        allOf:
        - $ref: '#/definitions/api.Pagination'
        description: page 信息
    type: object
...
  api.LocalDisksItemsList:
    properties:
      items:
        description: localDisks 节点磁盘列表
        items:
          $ref: '#/definitions/api.LocalDiskInfo'
        type: array
    type: object
...
  api.NodeUpdateReqBody:
    properties:
      enable:
        type: boolean
    type: object
  api.NodeUpdateRspBody:
    properties:
      success:
        type: boolean
    type: object
...
  api.Pagination:
    properties:
      page:
        description: 当前页索引，从 1 开始，为 0 时，会自动重置为默认值 constants.DefaultPage
        type: integer
      pageSize:
        description: |-
          每页数据量，为 -1 时表示查询全部，为 0 时会重置为默认值
          constants.DefaultPageSize
        type: integer
      pages:
        description: 总页数
        type: integer
      total:
        description: 总共有多少条目，请求时可以不用传递
        type: integer
    type: object
  api.RspFailBody:
    properties:
      description:
        type: string
      errcode:
        type: integer
    type: object
...
  api.StorageNode:
    properties:
      K8sNode:
        $ref: '#/definitions/v1.Node'
      k8SNodeState:
        $ref: '#/definitions/api.State'
      localDiskNode:
        $ref: '#/definitions/v1alpha1.LocalDiskNode'
      localStorageNode:
        $ref: '#/definitions/v1alpha1.LocalStorageNode'
    type: object
  api.StorageNodeList:
    properties:
      items:
        description: StorageNodes
        items:
          $ref: '#/definitions/api.StorageNode'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/api.Pagination'
        description: page 信息
    type: object
...
  api.StoragePool:
    properties:
      allocatedCapacityBytes:
        description: AllocatedCapacityBytes 存储池已经分配存储容量
        type: integer
      createTime:
        description: createTime 创建时间
        type: string
      items:
        items:
          $ref: '#/definitions/api.StorageNodePool'
        type: array
      nodeNames:
        description: NodesNames Pool所在节点列表
        items:
          type: string
        type: array
      poolName:
        description: 'Supported pool name: HDD_POOL, SSD_POOL, NVMe_POOL 存储池名称'
        type: string
      totalCapacityBytes:
        description: TotalCapacityBytes 存储池对应存储总容量
        type: integer
    type: object
...
  api.StoragePoolList:
    properties:
      items:
        description: storagePools
        items:
          $ref: '#/definitions/api.StoragePool'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/api.Pagination'
        description: page 信息
    type: object
...
  api.VolumeOperationListByNode:
    properties:
      items:
        description: VolumeOperations
        items:
          $ref: '#/definitions/api.VolumeMigrateOperation'
        type: array
      nodeName:
        description: node name
        type: string
      pagination:
        allOf:
        - $ref: '#/definitions/api.Pagination'
        description: page 信息
    type: object
...
  v1.Node:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        allOf:
        - $ref: '#/definitions/v1.ObjectMeta'
        description: |-
          Standard object's metadata.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
          +optional
      spec:
        allOf:
        - $ref: '#/definitions/v1.NodeSpec'
        description: |-
          Spec defines the behavior of a node.
          https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status
          +optional
      status:
        allOf:
        - $ref: '#/definitions/v1.NodeStatus'
        description: |-
          Most recently observed status of the node.
          Populated by the system.
          Read-only.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status
          +optional
    type: object
...
  v1alpha1.LocalDiskNode:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      spec:
        $ref: '#/definitions/v1alpha1.LocalDiskNodeSpec'
      status:
        $ref: '#/definitions/v1alpha1.LocalDiskNodeStatus'
    type: object
  v1alpha1.LocalDiskNodeSpec:
    properties:
      nodeName:
        description: NodeName represent where disk is attached
        type: string
    type: object
...
  v1alpha1.LocalDiskStatus:
    properties:
      claimState:
        allOf:
        - $ref: '#/definitions/v1alpha1.LocalDiskState'
        description: |-
          State represents the claim state of the disk
          +kubebuilder:validation:Enum:=Bound;Reserved;Available;Pending
    type: object
...
info:
  contact: {}
paths:
...
  /cluster/nodes:
    get:
      consumes:
      - application/json
      description: list StorageNode  驱动状态 [运行中（Ready）,维护中（Maintain）, 离线（Offline)]
        , 节点状态 [运行中（Ready）,未就绪（NotReady）,未知（Unknown)]
      parameters:
      - description: name
        in: query
        name: name
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        required: true
        type: integer
      - description: nodeState
        in: query
        name: nodeState
        type: string
      - description: driverState
        in: query
        name: driverState
        type: string
      - description: fuzzy
        in: query
        name: fuzzy
        type: boolean
      - description: sort
        in: query
        name: sort
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StorageNodeList'
      summary: 摘要 获取存储节点列表
      tags:
      - Node
  /cluster/nodes/{nodeName}:
    get:
      consumes:
      - application/json
      description: get StorageNode 驱动状态 [运行中（Ready）,维护中（Maintain）, 离线（Offline)] ,
        节点状态 [运行中（Ready）,未就绪（NotReady）,未知（Unknown)]
      parameters:
      - description: nodeName
        in: path
        name: nodeName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StorageNode'
      summary: 摘要 获取指定存储节点
      tags:
      - Node
    post:
      consumes:
      - application/json
      parameters:
      - description: nodeName
        in: path
        name: nodeName
        required: true
        type: string
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.NodeUpdateReqBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.NodeUpdateRspBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.RspFailBody'
      summary: Set node enable/disable
      tags:
      - Node
  /cluster/nodes/{nodeName}/disks:
    get:
      consumes:
      - application/json
      description: list StorageNodeDisks 状态枚举 （Active、Inactive、Unknown、Pending、Available、Bound)
      parameters:
      - description: nodeName
        in: path
        name: nodeName
        required: true
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        required: true
        type: integer
      - description: state
        in: query
        name: state
        type: string
      - description: fuzzy
        in: query
        name: fuzzy
        type: boolean
      - description: sort
        in: query
        name: sort
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LocalDiskListByNode'
      summary: 摘要 获取指定存储节点磁盘列表
      tags:
      - Node
  /cluster/nodes/{nodeName}/disks/{devicePath}:
    post:
      consumes:
      - application/json
      description: post UpdateStorageNodeDisk devicePath i.g /dev/sdb /dev/sdc ...
      parameters:
      - description: nodeName
        in: path
        name: nodeName
        required: true
        type: string
      - description: devicePath
        in: path
        name: devicePath
        required: true
        type: string
      - description: reqBody
        in: body
        name: body
        schema:
          $ref: '#/definitions/api.DiskReqBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.DiskReqBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.RspFailBody'
      summary: 摘要 更新磁盘
      tags:
      - Node
  /cluster/nodes/{nodeName}/disks/{diskName}:
    get:
      consumes:
      - application/json
      description: get GetStorageNodeDisk diskname i.g sdb sdc ...
      parameters:
      - description: nodeName
        in: path
        name: nodeName
        required: true
        type: string
      - description: diskName
        in: path
        name: diskName
        required: true
        type: string
      - description: fuzzy
        in: query
        name: fuzzy
        type: boolean
      - description: sort
        in: query
        name: sort
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LocalDiskInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.RspFailBody'
      summary: 摘要 获取指定磁盘信息
      tags:
      - Node
  /cluster/nodes/{nodeName}/migrates:
    get:
      consumes:
      - application/json
      description: get StorageNodeMigrate
      parameters:
      - description: nodeName
        in: path
        name: nodeName
        required: true
        type: string
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        required: true
        type: integer
      - description: migrateOperationName
        in: query
        name: migrateOperationName
        type: string
      - description: volumeName
        in: query
        name: volumeName
        type: string
      - description: state
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.VolumeOperationListByNode'
      summary: 摘要 获取指定节点数据卷迁移任务列表
      tags:
      - Node
  /cluster/nodes/{nodeName}/pools:
    get:
      consumes:
      - application/json
      description: get StorageNodePoolsList
      parameters:
      - description: nodeName
        in: path
        name: nodeName
        required: true
        type: string
      - description: fuzzy
        in: query
        name: fuzzy
        type: boolean
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: pageSize
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StoragePoolList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.RspFailBody'
      summary: 摘要 获取指定节点存储池列表信息
      tags:
      - Node
  /cluster/nodes/{nodeName}/pools/{poolName}:
    get:
      consumes:
      - application/json
      description: get StorageNodePoolGet
      parameters:
      - description: nodeName
        in: path
        name: nodeName
        required: true
        type: string
      - description: poolName
        in: path
        name: poolName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StoragePool'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.RspFailBody'
      summary: 摘要 获取指定节点指定存储池信息
      tags:
      - Node
  /cluster/nodes/{nodeName}/pools/{poolName}/disks:
    get:
      consumes:
      - application/json
      description: get StorageNodePoolDisksList
      parameters:
      - description: nodeName
        in: path
        name: nodeName
        required: true
        type: string
      - description: poolName
        in: path
        name: poolName
        required: true
        type: string
      - description: fuzzy
        in: query
        name: fuzzy
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LocalDisksItemsList'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.RspFailBody'
      summary: 摘要 获取指定节点指定存储池磁盘列表信息
      tags:
      - Node
  /cluster/nodes/{nodeName}/pools/{poolName}/disks/{diskName}:
    get:
      consumes:
      - application/json
      description: get StorageNodePoolDiskGet
      parameters:
      - description: nodeName
        in: path
        name: nodeName
        required: true
        type: string
      - description: poolName
        in: path
        name: poolName
        required: true
        type: string
      - description: diskName
        in: path
        name: diskName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LocalDiskInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.RspFailBody'
      summary: 摘要 获取指定节点指定存储池指定磁盘信息
      tags:
      - Node
...
swagger: "2.0"